Premium Analytics: move DataViewsDrilldownNative into the UI package
