+++ projects/packages/premium-analytics/changelog/add-pa-native-drilldown-table
@@ -1,4 +1,4 @@
 Significance: minor
 Type: added
 
-Widgets Toolkit: add the DataViewsDrilldownNative component (native DataViews showLevels hierarchy).
+UI: add the DataViewsDrilldownNative component (native DataViews showLevels hierarchy).
