+++ projects/packages/premium-analytics/packages/ui/src/dataviews-drilldown-native/__tests__/process-hierarchy-levels.test.ts
@@ -1,4 +1,4 @@
-import { processHierarchyLevels } from './process-hierarchy-levels';
+import { processHierarchyLevels } from '../process-hierarchy-levels';
 
 type Row = {
 	id: string;
