+++ projects/packages/premium-analytics/packages/widgets-toolkit/src/index.ts
@@ -27,9 +27,6 @@ export {
 	type LegendLabels,
 	LeaderboardLabel,
 	type LeaderboardLabelProps,
-	DataViewsDrilldownNative,
-	type DataViewsDrilldownNativeProps,
-	processHierarchyLevels,
 	BarChart,
 	type BarChartProps,
 	type BarChartData,
